Plant_Catalogue/requirements.txt
qrcode>=7.3.1
opencv-python>=4.8.0
pillow>=10.0.0
tabulate>=0.9.0
python-dateutil>=2.8.2
numpy>=1.24.0
configparser>=5.3.0
